+++ src/lib/Guiguts/KeyBindings.pm
@@ -114,7 +114,8 @@ sub keybindings {
     keybind( '<Control-Shift-z>', undef, '<<Redo>>' );    # Add another key-combination
 
     # Search
-    keybind( '<Control-f>', sub { ::searchpopup(); } );
+    keybind( '<Control-f>',       sub { ::searchpopup(); } );
+    keybind( '<Control-Shift-f>', sub { ::quicksearchpopup(); } );
     keybind(
         '<Control-g>',
         sub {
@@ -241,6 +242,9 @@ sub keybindings {
         keybind( '<Meta-x>',          sub { ::cut(); } );
         keybind( '<Meta-v>',          sub { ::paste(); } );
         keybind( '<Meta-f>',          sub { ::searchpopup(); } );
+        keybind( '<Meta-Shift-f>',    sub { ::quicksearchpopup(); } );
+        keybind( '<Meta-g>',          undef, '<<FindNext>>' );
+        keybind( '<Meta-Shift-g>',    undef, '<<FindNextReverse>>' );
         keybind( '<Meta-z>',          undef, '<<Undo>>' );
         keybind( '<Meta-y>',          undef, '<<Redo>>' );
         keybind( '<Meta-Up>',         [ 'SetCursor', '1.0' ] );
@@ -290,6 +294,8 @@ sub keybindings {
     # because default class bindings don't get set up until first Entry is created. So without
     # the dummy widget, the binding below would be overwritten by the default at a later point.
     $textwindow->MainWindow->bind( 'Tk::Entry', '<<Paste>>' => sub { ::entrypaste(shift); }, );
+    $textwindow->MainWindow->bind( 'Tk::Entry', '<Meta-v>'  => sub { ::entrypaste(shift); }, )
+      if $::OS_MAC;
 
     # Alternative paste to give user a second option if Perl/Tk utf8 bug strikes
     $textwindow->MainWindow->bind( 'Tk::Entry',